oracle-instance-damage-classification_corn/requirements.txt
torch
torchvision>=0.13
timm>=0.9.0
numpy
Pillow
PyYAML
matplotlib
scikit-learn
tqdm
